// src/interfaces/vendorInterfaces.ts
export interface VendorsData {
  address: string;
  contactEmail: string;
  contactName: string;
  contactPhoneNumber: string;
  description: string;
  website: string;
  name: string;
  isEnabled: boolean;
  id: number;
}

export type VendorFormInputs = {
  address: string;
  contactEmail: string;
  contactName: string;
  contactPhoneNumber: string;
  description: string;
  website: string;
  name: string;
};